refactor: Break "Add to Playlist" sheet into its own file
